Outbound packets can be shared
The packet contains mutable state regarding its position of writing to
the byte stream. This makes it impossible to share the same packet
to different connections.

This pr fixes that by introducing a flyweight that contains the
actual state, so that the original packets can be shared.
