+++ hazelcast/src/test/java/com/hazelcast/nio/serialization/PacketTest.java
@@ -16,27 +16,15 @@
 
 package com.hazelcast.nio.serialization;
 
-import com.hazelcast.internal.serialization.InternalSerializationService;
-import com.hazelcast.internal.serialization.SerializationServiceBuilder;
-import com.hazelcast.internal.serialization.impl.DefaultSerializationServiceBuilder;
 import com.hazelcast.nio.Packet;
-import com.hazelcast.spi.serialization.SerializationService;
 import com.hazelcast.test.HazelcastSerialClassRunner;
 import com.hazelcast.test.annotation.QuickTest;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
 import org.junit.runner.RunWith;
 
-import java.io.IOException;
-import java.nio.ByteBuffer;
-
 import static com.hazelcast.nio.Packet.FLAG_OP_CONTROL;
 import static com.hazelcast.nio.Packet.FLAG_URGENT;
-import static com.hazelcast.nio.serialization.SerializationConcurrencyTest.Address;
-import static com.hazelcast.nio.serialization.SerializationConcurrencyTest.FACTORY_ID;
-import static com.hazelcast.nio.serialization.SerializationConcurrencyTest.Person;
-import static com.hazelcast.nio.serialization.SerializationConcurrencyTest.PortableAddress;
-import static com.hazelcast.nio.serialization.SerializationConcurrencyTest.PortablePerson;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertSame;
@@ -45,56 +33,6 @@ import static org.junit.Assert.assertTrue;
 @RunWith(HazelcastSerialClassRunner.class)
 @Category(QuickTest.class)
 public class PacketTest {
-
-    private final Person person = new Person(111, 123L, 89.56d, "test-person", new Address("street", 987));
-
-    private final PortablePerson portablePerson = new PortablePerson(222, 456L, "portable-person",
-            new PortableAddress("street", 567));
-
-    private SerializationServiceBuilder createSerializationServiceBuilder() {
-        PortableFactory portableFactory = new PortableFactory() {
-            @Override
-            public Portable create(int classId) {
-                switch (classId) {
-                    case 1:
-                        return new PortablePerson();
-                    case 2:
-                        return new PortableAddress();
-                }
-                throw new IllegalArgumentException();
-            }
-        };
-        return new DefaultSerializationServiceBuilder().addPortableFactory(FACTORY_ID, portableFactory);
-    }
-
-    @Test
-    public void testPacketWriteRead() throws IOException {
-        testPacketWriteRead(person);
-    }
-
-    @Test
-    public void testPacketWriteRead_usingPortable() throws IOException {
-        testPacketWriteRead(portablePerson);
-    }
-
-    private void testPacketWriteRead(Object originalObject) throws IOException {
-        InternalSerializationService ss = (InternalSerializationService) createSerializationServiceBuilder().build();
-        byte[] originalPayload = ss.toBytes(originalObject);
-
-        ByteBuffer buffer = ByteBuffer.allocate(originalPayload.length * 2);
-        Packet originalPacket = new Packet(originalPayload);
-        assertTrue(originalPacket.writeTo(buffer));
-        buffer.flip();
-
-        SerializationService ss2 = createSerializationServiceBuilder().build();
-        Packet clonedPacket = new Packet();
-        assertTrue(clonedPacket.readFrom(buffer));
-
-        Object clonedObject = ss2.toObject(clonedPacket);
-
-        assertEquals(originalPacket, clonedPacket);
-        assertEquals(originalObject, clonedObject);
-    }
 
     @Test
     public void raiseFlags() {
